+++ ui/src/dashboards/components/SourceSelector.tsx
@@ -3,6 +3,7 @@ import React, {SFC} from 'react'
 
 // Componentes
 import SourceDropdown from 'src/flux/components/SourceDropdown'
+import {Radio} from 'src/reusable_ui'
 
 // Types
 import * as QueriesModels from 'src/types/queries'
@@ -14,8 +15,11 @@ interface Props {
   sources: SourcesModels.SourceOption[]
   service: Service
   services: Service[]
+  isFluxSource: boolean
+  sourceSupportsFlux: boolean
   queries: QueriesModels.QueryConfig[]
   isDynamicSourceSelected: boolean
+  toggleFlux: () => void
   onSelectDynamicSource: () => void
   onChangeService: (service: Service, source: SourcesModels.Source) => void
 }
@@ -26,7 +30,10 @@ const SourceSelector: SFC<Props> = ({
   service,
   services,
   queries,
+  toggleFlux,
+  isFluxSource,
   onChangeService,
+  sourceSupportsFlux,
   isDynamicSourceSelected,
   onSelectDynamicSource,
 }) => {
@@ -48,6 +55,29 @@ const SourceSelector: SFC<Props> = ({
         onChangeService={onChangeService}
         onSelectDynamicSource={onSelectDynamicSource}
       />
+      {isDynamicSourceSelected &&
+        sourceSupportsFlux && (
+          <Radio>
+            <Radio.Button
+              id="flux-source"
+              titleText="Flux"
+              value="Flux"
+              onClick={toggleFlux}
+              active={isFluxSource}
+            >
+              Flux
+            </Radio.Button>
+            <Radio.Button
+              id="influxql-source"
+              titleText="InfluxQL"
+              value="InfluxQL"
+              onClick={toggleFlux}
+              active={!isFluxSource}
+            >
+              InfluxQL
+            </Radio.Button>
+          </Radio>
+        )}
     </div>
   )
 }